requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
docutils==0.19
idna==3.4
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
numpy==1.23.4
Pygments==2.13.0
pypiwin32==223
python-tictactoe==0.0.7
pywin32==304
requests==2.28.1
urllib3==1.26.12
